// packages/api-client/src/generated/models/LogEntryDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LogEntryDto = {
    /**
     * Log identifier
     */
    id: string;
    /**
     * Log level
     */
    level: string;
    /**
     * Log message
     */
    message: string;
    /**
     * Structured context
     */
    context?: Record<string, any>;
    /**
     * Creation timestamp
     */
    createdAt: string;
};